Add the decompress function implementation
I've made two choices there. First we return the iolist and
nothing else. Second, we don't care about the size hint, or
the frame boundaries, and just decompress whatever the user
gave us even if there's multiple frames in there. We expect
the data to be one or more concatenated LZ4 frames.

The function is not perfect, especially with regard to
cleanup on errors. There's a few todos for that.