Makefile: Use base library's version and append a "revision"
This should make it clear what version of the lz4 library is being used.

The appended revision is similar to any packaging revisions. It useful
when we need to update the binding itself without changing the base
library.

We use a dot to separate the base library version and the revision
because that's what Erlang expects when comparing versions. See:
https://erlang.org/doc/system_principles/versions.html#version-scheme